+++ identity/app/controllers/EditProfileController.scala
@@ -5,7 +5,7 @@ import actions.AuthenticatedActions.AuthRequest
 import com.gu.identity.model.User
 import common.ImplicitControllerExecutionContext
 import form._
-import idapiclient.IdApiClient
+import idapiclient.{Error, IdApiClient}
 import model._
 import play.api.data.Form
 import play.api.i18n.{I18nSupport, MessagesApi, MessagesProvider}
@@ -123,7 +123,7 @@ case class ProfileForms(
     )
   }
 
-  def withErrors(errors: List[client.Error]): ProfileForms = {
+  def withErrors(errors: List[Error]): ProfileForms = {
     update{
       form =>
         errors.foldLeft(form){
