+++ identity/app/controllers/EmailController.scala
@@ -3,7 +3,7 @@ package controllers
 import actions.AuthenticatedActions
 import services.{IdRequestParser, IdentityUrlBuilder, ReturnUrlVerifier}
 import conf.IdentityConfiguration
-import idapiclient.IdApiClient
+import idapiclient.{Error, IdApiClient}
 import common.ImplicitControllerExecutionContext
 import utils.SafeLogging
 import play.api.mvc._
@@ -11,7 +11,6 @@ import play.api.mvc._
 import scala.concurrent.Future
 import model.{ApplicationContext, EmailNewsletters, IdentityPage}
 import play.api.data._
-import client.Error
 import com.gu.identity.model.{EmailList, Subscriber}
 import play.filters.csrf._
 import play.api.i18n.I18nSupport
