+++ identity/app/idapiclient/Error.scala
@@ -1,3 +1,3 @@
-package client
+package idapiclient
 
 case class Error(message: String, description: String, statusCode: Int = 500, context: Option[String] = None)
