+++ identity/app/idapiclient/HttpResponse.scala
@@ -1,3 +1,3 @@
-package client
+package idapiclient
 
 case class HttpResponse(body: String, statusCode: Int, statusMessage: String)
