+++ identity/app/idapiclient/package.scala
@@ -1,4 +1,4 @@
-package object client {
+package object idapiclient {
   type Response[T] = Either[List[Error], T]
   type Parameters = Iterable[(String, String)]
 }
