+++ identity/app/views/password/requestPasswordReset.scala.html
@@ -1,9 +1,9 @@
-@(page: model.Page, idRequest: services.IdentityRequest, idUrlBuilder: services.IdentityUrlBuilder, emailForm: Form[(String)], errors: List[client.Error])(implicit request: RequestHeader, messages: play.api.i18n.Messages, context: model.ApplicationContext)
+@(page: model.Page, idRequest: services.IdentityRequest, idUrlBuilder: services.IdentityUrlBuilder, emailForm: Form[(String)], errors: List[Error])(implicit request: RequestHeader, messages: play.api.i18n.Messages, context: model.ApplicationContext)
 
 @import form.IdFormHelpers._
+@import idapiclient.Error
 @import views.html.fragments.form.inputField
 @import views.html.fragments.registrationFooter
-@import client.Error
 
 @mainLegacy(page, projectName = Option("identity")){
 }{
