+++ identity/test/client/connection/dispatch/DispatchAsyncHttpClientTest.scala
@@ -1,11 +1,11 @@
 package client.connection.dispatch
 
-import client.HttpResponse
 import org.scalatest.path
 import org.scalatest.Matchers
 import org.scalatest.mockito.MockitoSugar
 import org.mockito.Mockito._
 import dispatch.Req
+import idapiclient.HttpResponse
 import test.WithTestExecutionContext
 
 import scala.concurrent.ExecutionContext
