+++ identity/test/client/parser/JsonBodyParserTest.scala
@@ -3,6 +3,8 @@ package client.parser
 import org.scalatest.path
 import net.liftweb.json.JsonAST.{JString, JValue}
 import client.{Error, HttpResponse}
+import idapiclient.{Error, HttpResponse}
+import idapiclient.parser.JsonBodyParser
 import org.scalatest.Matchers
 import net.liftweb.json.{DefaultFormats, Formats}
